# openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/memos/":{"get":{"tags":["Memos"],"summary":"Get Memos List","operationId":"get_memos_list_memos__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/MemoSchema"},"type":"array","title":"Response Get Memos List Memos  Get"}}}}}},"post":{"tags":["Memos"],"summary":"Create Memo","operationId":"create_memo_memos__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InsertAndUpdateMemoSchema"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseSchema"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/memos/{memo_id}":{"get":{"tags":["Memos"],"summary":"Get Memo Detail","operationId":"get_memo_detail_memos__memo_id__get","parameters":[{"name":"memo_id","in":"path","required":true,"schema":{"type":"integer","title":"Memo Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MemoSchema"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Memos"],"summary":"Modify Memo","operationId":"modify_memo_memos__memo_id__put","parameters":[{"name":"memo_id","in":"path","required":true,"schema":{"type":"integer","title":"Memo Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InsertAndUpdateMemoSchema"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseSchema"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Memos"],"summary":"Remove Memo","operationId":"remove_memo_memos__memo_id__delete","parameters":[{"name":"memo_id","in":"path","required":true,"schema":{"type":"integer","title":"Memo Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseSchema"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["Memos"],"summary":"Toggle Memo Completed","operationId":"toggle_memo_completed_memos__memo_id__patch","parameters":[{"name":"memo_id","in":"path","required":true,"schema":{"type":"integer","title":"Memo Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Body"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseSchema"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"InsertAndUpdateMemoSchema":{"properties":{"title":{"type":"string","minLength":1,"title":"Title","description":"メモのタイトルを入力してください。少なくとも1文字以上必要です。","example":"明日のアジェンダ"},"description":{"type":"string","title":"Description","description":"メモの内容についての追加情報。任意で記入できます。","default":"","example":"会議で話すトピック：プロジェクトの進捗状況"},"status":{"$ref":"#/components/schemas/MemoStatusSchema","description":"メモの状態を表す情報"}},"type":"object","required":["title","status"],"title":"InsertAndUpdateMemoSchema"},"MemoSchema":{"properties":{"title":{"type":"string","minLength":1,"title":"Title","description":"メモのタイトルを入力してください。少なくとも1文字以上必要です。","example":"明日のアジェンダ"},"description":{"type":"string","title":"Description","description":"メモの内容についての追加情報。任意で記入できます。","default":"","example":"会議で話すトピック：プロジェクトの進捗状況"},"status":{"$ref":"#/components/schemas/MemoStatusSchema","description":"メモの状態を表す情報"},"memo_id":{"type":"integer","title":"Memo Id","description":"メモを一意に識別するID番号。データベースで自動的に割り当てられます。","example":123},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Updated At"}},"type":"object","required":["title","status","memo_id","created_at"],"title":"MemoSchema"},"MemoStatusSchema":{"properties":{"priority":{"type":"string","title":"Priority","description":"優先度","example":"高"},"due_date":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Due Date","description":"メモの期限日、設定されていない場合はNone","example":"2023-10-01T00:00:00"},"is_completed":{"type":"boolean","title":"Is Completed","description":"メモが完了したかどうかを示すフラグ","default":false,"example":false}},"type":"object","required":["priority"],"title":"MemoStatusSchema"},"ResponseSchema":{"properties":{"message":{"type":"string","title":"Message","description":"API操作の結果を説明するメッセージ。","example":"メモの更新に成功しました。"}},"type":"object","required":["message"],"title":"ResponseSchema"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}